Handle `json-schema` URI schemes (#32)
This language server shows a Code Lens at the top of the file that you can
click on. It results in the client doing a workspace/executeCommand request.
When yaml-language-server receives that request, it in turn does a
window/showDocument request to the client. The URI of that window/showDocument
request has a `json-schema:` scheme. The fragment of that URI contains a HTTP
URL that you can use to fetch the JSON schema associated to the current YAML
file.